yolotube/server/requirements.txt
flask
google-cloud-storage
google-cloud-firestore
python-dotenv